Extract typeExtractor to its own file
